requirements.txt
###### Requirements without Version Specifiers ######`
# function tools
pyqt5
pyqt-darktheme
yongshi-pynfc
sqlitemanager
ndef
pyscard

# buildtools
setuptools
twine
pyinstaller

###### Requirements with Version Specifiers ######`
# pyqt5 >= 0.6.1             # Version Matching. Must be version 0.6.1